requirements.txt
# ============================================================
# requirements.txt — ts-forecast-project
# Toutes les dépendances versionnées pour reproductibilité
# Installer : pip install -r requirements.txt
# ============================================================

# --- data collection ---
yfinance
openmeteo-requests
httpx

# --- data engineering ---
pandas
numpy
matplotlib
seaborn
pyarrow
scipy

# --- utilities ---
pyyaml
python-dotenv
joblib
tqdm
statsmodels

# --- Jupyter ---
jupyter
jupyterlab
nbconvert
ipywidgets

# --- MLflow ---
mlflow

# --- model ---
xgboost
scikit-learn
torch
prophet
optuna

# --- monitoring & drift detection ---
evidently
prometheus-client

# --- versioning & deployment ---
dvc
dvc-gdrive
fastapi
uvicorn[standard]
pydantic
flask

# --- testing ---
pytest
pytest-cov
pytest-asyncio